requirements.txt
affine
aniso8601
annotated-types
anyio
asttokens
attrs
bleach
blinker
bokeh
boto3
botocore
certifi
cftime
charset-normalizer
click
click-plugins
cligj
colorcet
comm
contourpy
cycler
debugpy
decorator
distro
docutils
executing
feedgenerator
Flask
Flask-Cors
Flask-RESTful
fonttools
gunicorn
h11
httpcore
httpx
idna
imageio
ipykernel
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
jiter
jmespath
jsonpatch
jsonpointer
jupyter_client
jupyter_core
kiwisolver
langchain
langchain-core
langchain-openai
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
linkify-it-py
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
narwhals
nest-asyncio
netCDF4
numpy
openai
opencv-python
OpenVisus
openvisuspy
ordered-set
orjson
ormsgpack
packaging
pandas
panel
param
parso
pelican
pelican-data
pexpect
pillow
platformdirs
pooch
prompt_toolkit
psutil
ptyprocess
pure_eval
pybind11
pydantic
pydantic_core
Pygments
pyparsing
PyQt5
PyQt5-Qt5
PyQt5_sip
python-dateutil
python-dotenv
pytz
pyvista
pyviz_comms
PyYAML
pyzmq
rasterio
regex
requests
requests-toolbelt
rich
s3transfer
scipy
scooby
setuptools
six
sniffio
stack-data
tenacity
tiktoken
tornado
tqdm
traitlets
typing-inspection
typing_extensions
tzdata
uc-micro-py
Unidecode
urllib3
vtk
watchfiles
wcwidth
webencodings
Werkzeug
xarray
xmltodict
xxhash
xyzservices
zstandard
